Changes on request 1) removed grid lines between JSH and non-JSH data 2) Jockey Rs, FP and Trainer Rs column in Bold 3) Mov9-60 column when less than 0 display in red font when greater than 0 display in bold 4) Changed Mov9-60 column to an integer

+++ src/main/java/uk/co/punishell/insideview/view/commands/entityCommands/JSHRunnerCommand.java
@@ -43,5 +43,5 @@ public class JSHRunnerCommand {
     private int trainerRunnerNo;
     private int trainerRunners;
     private String headGear;
-    private double mov9to60;
+    private int mov9to60;
 }
